requirements.txt
requests-futures
requests
pychalk
asyncio
python-nmap
bs4